add flake8 to user, kitchen_recipes

+++ webapp/kitchen_recipes/views.py
@@ -1,13 +1,14 @@
-from flask import Blueprint, flash, render_template, redirect, url_for, abort
+from flask import Blueprint, abort, flash, redirect, render_template, url_for
+
 from flask_login import current_user
 
-from webapp.kitchen_recipes.forms import AddNewRecipeForm
-from webapp.kitchen_recipes.models import Recipe, Photo
-from webapp.services.service_photo import is_extension_allowed, save_files
 from webapp.db import db
-
+from webapp.kitchen_recipes.forms import AddNewRecipeForm
+from webapp.kitchen_recipes.models import Photo, Recipe
+from webapp.services.service_photo import is_extension_allowed, save_files
 
 blueprint = Blueprint('kitchen_recipes', __name__)
+
 
 @blueprint.route('/')
 def index():
@@ -18,8 +19,7 @@ def index():
         'kitchen_recipes/index.html',
         title=title,
         context=context,
-        recipes=recipes
-    )
+        recipes=recipes)
 
 
 @blueprint.route('/<int:recipe_id>')
@@ -29,8 +29,7 @@ def page_recipe(recipe_id):
         abort(404)
     return render_template(
         'kitchen_recipes/page_recipe.html',
-        recipe=recipe
-    )    
+        recipe=recipe)
 
 
 @blueprint.route('/add_recipe')
@@ -40,8 +39,7 @@ def add_recipe():
     return render_template(
         'kitchen_recipes/add_recipe.html',
         title=title,
-        form_add_recipe=form_add_recipe
-    )
+        form_add_recipe=form_add_recipe)
 
 
 @blueprint.route('/process_add_recipe', methods=['POST'])
@@ -51,40 +49,35 @@ def process_add_recipe():
 
         photo = form.photo.data
 
-        if is_extension_allowed(photo) == False:
+        if is_extension_allowed(photo) is False:
             flash('Можно добавить изображения с расширеним png, jpg, jpeg')
             return redirect(url_for('marketplace.add_product'))
 
         photo_path = save_files(photo)
 
         new_recipe = Recipe(
-            category_id = form.category.data,
+            category_id=form.category.data,
             user_id=current_user.id,
-            name = form.name.data,
-            description = form.description.data,
+            name=form.name.data,
+            description=form.description.data,
         )
-    
         db.session.add(new_recipe)
         db.session.commit()
 
         new_recipe_photo = Photo(
             recipe_id=new_recipe.id,
-            photo_path=photo_path
+            photo_path=photo_path,
         )
 
         db.session.add(new_recipe_photo)
         db.session.commit()
-
-        
-    
         flash('Вы добавили рецепт')
         return redirect(url_for('kitchen_recipes.index'))
     else:
         for field, error in form.errors.items():
             flash('Ошибка в поле {}: {}'.format(
                 getattr(form, field).lavel.text,
-                error
-            ))
+                error))
     return redirect(url_for('kitchen_recipes.add_recipe'))
 
 
@@ -93,9 +86,7 @@ def process_delete_recipe(recipe_id):
     if not current_user.is_authenticated:
         return redirect(url_for('kitchen_recipes.index'))
     else:
-        delete_recipe = Recipe.query.filter(Recipe.user_id == current_user.id,
-                                            Recipe.id == recipe_id
-                                    ).delete()
+        Recipe.query.filter(Recipe.user_id == current_user.id, Recipe.id == recipe_id).delete()
         db.session.commit()
         flash('Вы успешно удалили рецепт')
         return redirect(url_for('kitchen_recipes.index'))
